+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant.Test/Plan/PlanMocks.cs
@@ -81,7 +81,7 @@ namespace NINA.Plugin.Assistant.Test.Plan {
             pe.SetupProperty(m => m.Desired, desired);
             pe.SetupProperty(m => m.Acquired, 0);
             pe.SetupProperty(m => m.Accepted, accepted);
-            pe.Setup(m => m.NeededExposures()).Returns(accepted > desired ? 0 : desired - accepted);
+            pe.Setup(m => m.NeededExposures(125)).Returns(accepted > desired ? 0 : desired - accepted);
             pe.Setup(m => m.IsIncomplete()).Returns(accepted < desired);
             return pe;
         }