+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Database/Migrate/SQL.Designer.cs
@@ -116,5 +116,22 @@ namespace Assistant.NINAPlugin.Database.Migrate {
                 return ResourceManager.GetString("3", resourceCulture);
             }
         }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to /*
+        ///*/
+        ///
+        ///ALTER TABLE profilepreference ADD COLUMN acceptimprovement INTEGER DEFAULT 1;
+        ///ALTER TABLE profilepreference ADD COLUMN exposurethrottle REAL DEFAULT 100;
+        ///ALTER TABLE profilepreference ADD COLUMN parkonwait INTEGER DEFAULT 0;
+        ///
+        ///PRAGMA user_version = 4;
+        ///.
+        /// </summary>
+        internal static string _4 {
+            get {
+                return ResourceManager.GetString("4", resourceCulture);
+            }
+        }
     }
 }