+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Database/SchedulerPlanLoader.cs
@@ -9,9 +9,27 @@ namespace Assistant.NINAPlugin.Database {
 
     public class SchedulerPlanLoader {
 
-        public List<IPlanProject> LoadActiveProjects(SchedulerDatabaseContext context, IProfile activeProfile) {
+        private IProfile activeProfile;
+        private string profileId;
+
+        public SchedulerPlanLoader(IProfile activeProfile) {
+            this.activeProfile = activeProfile;
+            profileId = activeProfile.Id.ToString();
+        }
+
+        public ProfilePreference GetProfilePreferences(SchedulerDatabaseContext context) {
+            using (context) {
+                ProfilePreference profilePreference = context.GetProfilePreference(profileId);
+                if (profilePreference == null) {
+                    profilePreference = new ProfilePreference(profileId);
+                }
+
+                return profilePreference;
+            }
+        }
+
+        public List<IPlanProject> LoadActiveProjects(SchedulerDatabaseContext context) {
             List<Project> projects = null;
-            string profileId = activeProfile.Id.ToString();
 
             using (context) {
                 try {